wgml: do code more portable
remove/change 16-bit support residual code and data

+++ bld/wgml1/c/gsli.c
@@ -87,7 +87,7 @@ void    scr_li( void )
             case 2 :
                 if( strnicmp( "ON", pa, 2 ) == 0 ) {
                     if( !ProcFlags.literal ) {
-                        li_cnt = LONG_MAX;
+                        li_cnt = INT_MAX;
                         ProcFlags.literal = true;
                         scan_restart = pa + 2;
                     }
